backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
opencv-python==4.8.1.78
numpy==1.24.3
ultralytics==8.0.200
torch==2.1.0
torchvision==0.16.0
python-socketio==5.10.0
websockets==12.0
pillow==10.1.0
scipy==1.11.3
